LPS-72816 Moves classes
