+++ modules/apps/web-experience/layout/layout-type-controller/layout-type-controller-test/src/testIntegration/java/com/liferay/layout/type/controller/test/LayoutTypeControllerTrackerTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.util;
+package com.liferay.layout.type.controller.test;
 
+import com.liferay.layout.type.controller.TestLayoutTypeController;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.LayoutTypeController;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -22,7 +23,7 @@ import com.liferay.portal.model.impl.LayoutImpl;
 import com.liferay.portal.model.impl.LayoutTypeControllerImpl;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.SyntheticBundleRule;
-import com.liferay.portal.util.bundle.layouttypecontrollertracker.TestLayoutTypeController;
+import com.liferay.portal.util.LayoutTypeControllerTracker;
 
 import java.util.Locale;
 import java.util.Map;
