--- portal-impl/test/integration/com/liferay/portal/util/bundle/layouttypecontrollertracker/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Description: Test Bundle for LayoutTypeControllerTrackerTest
-Bundle-Version: 1.0.0
-Private-Package: $(bundle.package).*
--dsannotations: *
